+++ src/Adapter/Server/SERVER_CONFIG.mjs
@@ -1,0 +1,17 @@
+export const SERVER_CONFIG_HTTPS_CERTIFICATE_KEY = "server-https-certificate";
+
+export const SERVER_CONFIG_HTTPS_DHPARAM_KEY = "server-https-dhparam";
+
+export const SERVER_CONFIG_HTTPS_KEY_KEY = "server-https-key";
+
+export const SERVER_CONFIG_LISTEN_HTTP_PORT_KEY = "server-listen-http-port";
+
+export const SERVER_CONFIG_LISTEN_HTTPS_PORT_KEY = "server-listen-https-port";
+
+export const SERVER_CONFIG_LISTEN_INTERFACE_KEY = "server-listen-interface";
+
+export const SERVER_CONFIG_REDIRECT_HTTP_TO_HTTPS_KEY = "server-redirect-http-to-https";
+
+export const SERVER_CONFIG_REDIRECT_HTTP_TO_HTTPS_PORT_KEY = "server-redirect-http-to-https-port";
+
+export const SERVER_CONFIG_REDIRECT_HTTP_TO_HTTPS_STATUS_CODE_KEY = "server-redirect-http-to-https-status-code";